Refactoring search.js
