--- favorites.html
@@ -1,21 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Favoritos</title>
-  <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@500;700&family=Roboto:wght@400;700&display=swap"
-    rel="stylesheet">
-  <link rel="stylesheet" href="styles/normalize.css">
-  <link rel="stylesheet" href="styles/dark-mode.css">
-  <link rel="stylesheet" href="styles/favorites.css">
-  <link rel="stylesheet" href="icomoon/style.css">
-</head>
-
-<body>
-  <main class="main">
-  
-  </main>
-  <script src="scripts/favorites.js" type="module"></script>
-</body>
-</html>
