+++ index.html
@@ -13,6 +13,7 @@
   <link rel="stylesheet" href="styles/hamburguer_menu.css">
   <link rel="stylesheet" href="styles/dark-mode.css">
   <link rel="stylesheet" href="styles/desktop.css">
+  <link rel="stylesheet" href="styles/favorites.css">
   <link rel="stylesheet" href="styles/search.css">
 </head>
